requirements.txt
absl-py==0.9.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-auth==1.11.2
google-auth-oauthlib==0.4.1
grpcio==1.27.2
idna==2.9
Markdown==3.2.1
numpy==1.18.1
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
tensorboard==2.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
urllib3==1.25.8
Werkzeug==1.0.0
yapf==0.29.0
